refactor: remove single file tests
