--- packages/backend/test/auth.getIdentity.test.ts
@@ -1,8 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-describe('api.auth.getIdentity', () => {
-  it('TODO', () => {
-    // TODO
-    expect(true).toBe(true);
-  });
-});
