--- packages/backend/test/data_ml/friendRecs.test.ts
@@ -1,8 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-describe('api["data_ml/friendRecs"]', () => {
-  it('upsert', () => {
-    // TODO
-    expect(true).toBe(true);
-  });
-});
