+++ packages/backend/test/data_ml/friends.test.ts
@@ -1,13 +1,17 @@
 import { describe, expect, it } from 'vitest';
 
 describe('api["data_ml/friends"]', () => {
-  it('friendExists', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('friendExists', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 
-  it('getFriendRecs', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('getFriendRecs', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 });
