+++ packages/backend/test/data_ml/universal.test.ts
@@ -1,8 +1,10 @@
 import { describe, expect, it } from 'vitest';
 
 describe('api["data_ml/universal"]', () => {
-  it('queryAll', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('queryAll', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 });
