+++ packages/backend/test/data_ml/updateUserPreferences.test.ts
@@ -1,18 +1,24 @@
 import { describe, expect, it } from 'vitest';
 
 describe('api["data_ml/updateUserPreferences"]', () => {
-  it('getByUserId', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('getByUserId', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 
-  it('getByEventId', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('getByEventId', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 
-  it('upsertUserTagWeights', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('upsertUserTagWeights', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 });
