+++ packages/backend/test/data_ml/users.test.ts
@@ -1,18 +1,24 @@
 import { describe, expect, it } from 'vitest';
 
 describe('api["data_ml/users"]', () => {
-  it('queryByToken', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('queryByToken', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 
-  it('userExists', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('userExists', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 
-  it('getUserIdsWithEventAttendance', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('getUserIdsWithEventAttendance', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 });
