--- packages/backend/test/data_ml/users.userExists.test.ts
@@ -1,8 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-describe('api["data_ml/users"].userExists', () => {
-  it('TODO', () => {
-    // TODO
-    expect(true).toBe(true);
-  });
-});
