--- packages/backend/test/eventsIngest.syncTicketmasterLasVegas.test.ts
@@ -1,8 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-describe('api.eventsIngest.syncTicketmasterLasVegas', () => {
-  it('TODO', () => {
-    // TODO
-    expect(true).toBe(true);
-  });
-});
