--- packages/backend/test/seed.seed.test.ts
@@ -1,8 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-describe('api.seed.seed', () => {
-  it('TODO', () => {
-    // TODO
-    expect(true).toBe(true);
-  });
-});
