+++ packages/backend/test/seed.test.ts
@@ -1,8 +1,10 @@
 import { describe, expect, it } from 'vitest';
 
 describe('api.seed', () => {
-  it('seed', () => {
-    // TODO
-    expect(true).toBe(true);
+  describe('seed', () => {
+    it('TODO', () => {
+      // TODO
+      expect(true).toBe(true);
+    });
   });
 });
