--- packages/backend/test/users.getProfileByName.test.ts
@@ -1,8 +1,0 @@
-import { describe, expect, it } from 'vitest';
-
-describe('api.users.getProfileByName', () => {
-  it('TODO', () => {
-    // TODO
-    expect(true).toBe(true);
-  });
-});
